Add test initialisation
