+++ vm_server/send/proto/Request.proto
@@ -45,5 +45,3 @@ message TaskStatusResponse {
   Status status = 2;                                                                                               
   TaskResponse task_response = 3; 
   }
-
-
